+++ tests/v0_tests/test_boost_search.py
@@ -15,12 +15,13 @@ class TestBoostSearch(MarqoTestCase):
     def setUp(self) -> None:
         self.client = Client(**self.client_settings)
         self.index_name_1 = "my-test-index-1"
-        try:
-            self.client.delete_index(self.index_name_1)
-        except MarqoApiError as s:
-            pass
-        self.client.create_index(index_name = self.index_name_1)
-        self.client.index(index_name = self.index_name_1).add_documents(
+        if not self.client.config.is_marqo_cloud:
+            try:
+                self.client.delete_index(self.index_name_1)
+            except MarqoApiError as s:
+                pass
+        self.create_index(index_name=self.index_name_1)
+        self.client.index(index_name=self.index_name_1).add_documents(
             [
                 {
                     "Title": "A comparison of the best pets",
@@ -39,11 +40,13 @@ class TestBoostSearch(MarqoTestCase):
         self.query = "What are the best pets"
 
     def tearDown(self) -> None:
-        try:
-            self.client.delete_index(self.index_name_1)
-        except MarqoApiError as s:
-            pass
-
+        if not self.client.config.is_marqo_cloud:
+            try:
+                self.client.delete_index(self.index_name_1)
+            except MarqoApiError as s:
+                pass
+        else:
+            self.delete_documents(self.index_name_1)
 
     def test_boost_search_format(self):
         if self.IS_MULTI_INSTANCE:
