+++ tests/v0_tests/test_client.py
@@ -2,6 +2,7 @@ import requests
 from json import JSONDecodeError
 from unittest import mock
 from unittest.mock import patch
+from pytest import mark
 
 from marqo.client import Client
 from tests.marqo_test import MarqoTestCase
@@ -14,26 +15,33 @@ class TestClient(MarqoTestCase):
     def setUp(self) -> None:
         self.client = Client(**self.client_settings)
         self.index_name_1 = "my-test-index-1"
-        try:
-            self.client.delete_index(self.index_name_1)
-        except MarqoApiError as s:
-            pass
+        if not self.client.config.is_marqo_cloud:
+            try:
+                self.client.delete_index(self.index_name_1)
+            except MarqoApiError as s:
+                pass
 
     def tearDown(self) -> None:
-        try:
-            self.client.delete_index(self.index_name_1)
-        except MarqoApiError as s:
-            pass
+        if not self.client.config.is_marqo_cloud:
+            try:
+                self.client.delete_index(self.index_name_1)
+            except MarqoApiError as s:
+                pass
+        else:
+            self.delete_documents(self.index_name_1)
 
+    @mark.ignore_cloud_tests
     def test_get_marqo(self):
         res = self.client.get_marqo()
         assert 'Welcome to Marqo' == res['message']
 
+    @mark.ignore_cloud_tests
     def test_health(self):
         res = self.client.health()
         assert 'status' in res
         assert 'status' in res['backend']
 
+    @mark.ignore_cloud_tests
     def test_health_deprecation_warning(self):
         with mock.patch("marqo.client.mq_logger.warning") as mock_warning:
             res = self.client.health()
@@ -43,6 +51,7 @@ class TestClient(MarqoTestCase):
             warning_message = mock_warning.call_args[0][0]
             self.assertIn("The `client.health()` API has been deprecated and will be removed in", warning_message)
 
+    @mark.ignore_cloud_tests
     def test_error_handling_in_health_check(self):
         client = Client(**self.client_settings)
         side_effect_list = [requests.exceptions.JSONDecodeError("test", "test", 1), BackendCommunicationError("test"),
@@ -59,14 +68,14 @@ class TestClient(MarqoTestCase):
                 self.assertIn("If you are trying to check the health on Marqo Cloud", cm.exception.message)
 
     def test_check_index_health_response(self):
-        self.client.create_index(self.index_name_1)
+        self.create_index(self.index_name_1)
         res = self.client.index(self.index_name_1).health()
         assert 'status' in res
         assert 'status' in res['backend']
 
     def test_check_index_health_query(self):
         with patch("marqo._httprequests.HttpRequests.get") as mock_get:
-            self.client.create_index(self.index_name_1)
+            self.create_index(self.index_name_1)
             res = self.client.index(self.index_name_1).health()
             args, kwargs = mock_get.call_args
             self.assertIn(f"/{self.index_name_1}/health", kwargs["path"])
